+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The Branch Web SDK adheres to [Semantic Versioning](http://semver.org/).
 - Specify whether the app banner is "sticky" (position: fixed) on desktop and mobile
 - Set the app banner to either never show a second time, or wait a specified number of days before showing again
 - Adds `customCSS` property to banner to allow custom styles
+- Adds previously Cordova only functions to Web: `creditHistory()`, `getCode()`, `validateCode()`, and `applyCode()`
 
 ## [v1.5.1] - 2015-05-13
 - Fixes IE11 bug with XHRRequests that caused an InvalidState error